requirements.txt
asgiref==3.5.2
autopep8==2.0.0
cffi==1.15.1
cryptography==38.0.3
Django==4.1.3
django-cors-headers==3.13.0
djangorestframework==3.14.0
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.2.2
mysql==0.0.3
mysqlclient==2.1.1
pycodestyle==2.9.1
pycparser==2.21
PyJWT==1.7.1
PyMySQL==1.0.2
pytz==2022.6
sqlparse==0.4.3
tomli==2.0.1
